Broker divided into two classes
