+++ wallet-console/src/main/java/com/isa/jjdzr/WalletApp.java
@@ -1,6 +1,6 @@
 package com.isa.jjdzr;
 
-import com.isa.jjdzr.console.Menu;
+import com.isa.jjdzr.consolepresentation.Menu;
 
 public class WalletApp
 {
