--- wallet-console/src/main/java/com/isa/jjdzr/console/App.java
@@ -1,10 +1,0 @@
-package com.isa.jjdzr.console;
-
-import java.io.IOException;
-
-public class App {
-    public static void main(String[] args)  {
-        Printable menu = new Menu2();
-        menu.printIncomindCash("100");
-    }
-}
