+++ wallet-console/src/main/java/com/isa/jjdzr/consolepresentation/AssetsViewer.java
@@ -5,7 +5,6 @@ import com.isa.jjdzr.dto.WalletAsset;
 import com.isa.jjdzr.market.Market;
 
 import java.math.BigDecimal;
-import java.sql.SQLOutput;
 import java.util.ArrayList;
 import java.util.List;
