+++ wallet-console/src/main/java/com/isa/jjdzr/consolepresentation/BrokerCreator.java
@@ -3,6 +3,6 @@ package com.isa.jjdzr.consolepresentation;
 import com.isa.jjdzr.dto.Wallet;
 
 public interface BrokerCreator {
-    Broker createBroker();
-    Broker createBrokerWithWallet(Wallet wallet);
+    BrokerBuy createBroker();
+    BrokerBuy createBrokerWithWallet(Wallet wallet);
 }
