+++ wallet-console/src/main/java/com/isa/jjdzr/consolepresentation/LoadViewer.java
@@ -1,8 +1,7 @@
 package com.isa.jjdzr.consolepresentation;
 
-import com.isa.jjdzr.console.Menu;
-import com.isa.jjdzr.console.Menu2;
 import com.isa.jjdzr.console.Printable;
+import com.isa.jjdzr.console.Printer;
 import com.isa.jjdzr.datareadandwrite.WalletLoader;
 import com.isa.jjdzr.dto.Wallet;
 
@@ -10,11 +9,11 @@ import java.util.Scanner;
 import java.util.regex.Pattern;
 
 public class LoadViewer {
-    private Printable menu = new Menu2();
+    private Printable printer = new Printer();
     public Wallet load() {
         Scanner scan = new Scanner(System.in);
         Wallet wallet = null;
-        menu.printActualLine("Za chwilę nastąpi wczytanie portfela.");
+        printer.printActualLine("Za chwilę nastąpi wczytanie portfela.");
         if (doYouWantToContinue()) {
             wallet = new WalletLoader().loadWallet();
             if (wallet == null) {
